requirements_analyst.txt
python-dotenv
selenium
webdriver-manager
google-generativeai
openai
Pillow
requests
